requirements.txt
tensorflow-io-gcs-filesystem==0.31.0
tensorflow==2.19.0
tensorflow-hub==0.16.1
pillow==11.1.0
lightgbm==4.6.0
pandas==2.2.3
matplotlib==3.10.1
seaborn==0.13.2
scikit-learn==1.6.1
ipykernel==6.29.5